snowpark/pyproject.toml
[tool.poetry]
name = "snowpark"
version = "0.1.0"
description = ""
authors = ["Arthur Ribeiro"]

[tool.poetry.dependencies]
python = "~3.8"
loguru = "^0.7.0"
snowflake-connector-python = "^3.0.4"
snowflake-snowpark-python = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
